Refactor order and trade processing for flexibility.
Updated the `GetOpenedOrdersFromOneInitialSignal` method to use `long` type for `messageId` and adjusted order status checks. Introduced `GetTradesByInitialSignalMessageId` in `TradeRepository` for enhanced trade retrieval based on initial signals. Added `ServiceOrderRemoveOrderProcessor` for processing order cancellation actions.

+++ MetaTraderWorkerService/Repository/Orders/IOrderRepository.cs
@@ -11,5 +11,5 @@ public interface IOrderRepository
     Task<List<MetaTraderOrder>> GetPlacedOrdersAsync();
     Task<MetaTraderOrder> GetOrderByMagicAndSymbolAsync(int tradeMagic, string tradeSymbol);
     Task<MetaTraderOrder> GetOrderByMetaTraderOrderId(string metaTraderOrderId, string tradeSymbol);
-    Task<List<MetaTraderOrder>> GetOpenedOrdersFromOneInitialSignal(int messageId);
+    Task<List<MetaTraderOrder>> GetOpenedOrdersFromOneInitialSignal(long messageId);
 }
